Editorial pass
We didn't really say anything about rekeying, which is the most obvious
way of managing limits.

This was raised (indirectly) by @emanjon (via #37) and also @smyshsv in #10.

I considered doing more for #10 here, but I think that it is probably
best to leave the RFC 8645 AEAD definitions (those with internal
rekeying) out of this.  We've very deliberately concentrated on modes
that have widespread deployment and -- cool as they are -- I am not
aware of any implementation of internal rekeying modes.  (FWIW, it looks
like the analysis there is entirely consistent with the multi-key
setting results we have.)

Closes #10.