Simplify ChaCha limits
The extra term really isn't relevant here unless you have enormous
messages.  By which I mean, messages that are 32 Yottabytes in size.
That's a single encryption of that size.  2^100 of those messages, with
every one of them that big.  No rekeying either.

It's a little less implausble if q << 2^100.  A q of 1 results in a
message size of 2^139 bytes.

Neither is concretely possible as RFC 7539 has just a 32-bit counter,
which limits messages to 256GB.