Convert all instances of l to L
This was pretty hard to automate as you might imagine.

I also tried to be more consistent in the application of backticks
throughout.  There are a few changes that relate to that.  I've also
tried to remove a few places where dropping a variable name into prose
might be jarring.