AE limits are what matter most
We use CA and IA as steps toward placing a bound on AEA;
we shouldn't emphasize the use of these things separately.

As Thomas notes, your AEAD doesn't work well if you lose integrity
(there are modes that can do that catastrophically, even).
More importantly, IND-CPA isn't much chop against an attacker that can
apply arbitrary forgeries because integrity is busted.
But this is an AEAD, with both parts active.  As long as both pieces are
functional, we retain confidentiality.

This attempts to make a few targeted choices to properly emphasize that
point.